security updates

+++ ds2024Team77/src/main/java/gr/hua/dit/ds/ds2024Team77/controllers/ProjectsController.java
@@ -9,6 +9,7 @@ import gr.hua.dit.ds.ds2024Team77.service.UserDetailsImpl;
 import gr.hua.dit.ds.ds2024Team77.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
@@ -40,6 +41,8 @@ public class ProjectsController {
         return pService.getProjects();
     }
 
+
+    @Secured({"ROLE_BASIC"})
     @PostMapping("")
     public void createProject(@RequestBody Project project,
                                  @AuthenticationPrincipal UserDetailsImpl auth){
@@ -49,15 +52,29 @@ public class ProjectsController {
         pService.saveProject(project);
     }
 
+    @Secured({"ROLE_BASIC"})
     @GetMapping("/{projectId}")
     public ResponseEntity<Project> getProject(@PathVariable Long projectId){
         Optional<Project> project = pService.getProject(projectId);
         return project.map(ResponseEntity::ok).orElseGet(()->ResponseEntity.notFound().build());
     }
 
+    @Secured({"ROLE_BASIC"})
     @DeleteMapping("/{projectId}")
-    public ResponseEntity<String> deleteProject(@PathVariable Long projectId){
+    public ResponseEntity<String> deleteProject(@PathVariable Long projectId,
+                                                @AuthenticationPrincipal UserDetailsImpl auth){
 
+        Optional<Project> projectOptional = pService.getProject(projectId);
+
+        if (projectOptional.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Project not found.");
+        }
+
+        Project project = projectOptional.get();
+
+        if (!project.getCustomer().getId().equals(auth.getId())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("You are not authorized to delete this project.");
+        }
         boolean result = this.pService.deleteProjectById(projectId);
         if(result){
             return ResponseEntity.status(HttpStatus.OK).body("Project deleted successfully.");
@@ -82,14 +99,21 @@ public class ProjectsController {
         return pService.getProjectsByFreelancer_Id(auth.getId());
     }
 
+    @Secured("ROLE_ADMIN")
     @PutMapping("/{projectId}/approve")
     public ResponseEntity<String> approveProject(@PathVariable Long projectId) {
-        boolean result = pService.approveProject(projectId);
-        if (result) {
-            return ResponseEntity.status(HttpStatus.OK).body("Project approved successfully.");
-        } else {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Project not found.");
+
+        try{
+            boolean result = pService.approveProject(projectId);
+            if (result) {
+                return ResponseEntity.status(HttpStatus.OK).body("Project approved successfully.");
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Project not found.");
+            }
+        }catch (Exception e){
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error approving project: " + e.getMessage());
         }
+
     }
 
 }
